+++ lesson_01/mvc_framework/templator.py
@@ -1,4 +1,3 @@
-import pathlib
 from jinja2 import Template
 from pathlib import Path
 
@@ -20,6 +19,6 @@ if __name__ == '__main__':
     res = render_from_line('Hello {{ name }}', context={'name': 'John Doe'})
     print(res)
 
-    template_file = Path.cwd() / 'lesson_01' / 'stub.html'
+    template_file = Path.cwd() / 'templates' / 'stub.html'
     res = render(template_file, context={'name': 'John Doe'})
     print(res)
