+++ lesson_01/mvc_framework/urls.py
@@ -1,0 +1,6 @@
+from .view import index, Info
+
+routes: dict[str, callable] = {
+    '/': index,
+    '/info': Info(),
+}
